format some stuff

+++ Weaver/Registry/VariableRegistry.cs
@@ -92,6 +92,7 @@ namespace Weaver.Registry
                     {
                         _store[existingRange.Start + i] = elements[i];
                     }
+
                     return; // We are done. No allocation needed!
                 }
 
@@ -137,6 +138,7 @@ namespace Weaver.Registry
                         _store[index] = kvp.Value.WithAttribute(kvp.Key);
                         index++;
                     }
+
                     return; // Done without allocating!
                 }
 
@@ -186,7 +188,7 @@ namespace Weaver.Registry
             // you would need to change private readonly Dictionary<int, VmValue> _store 
             // to Dictionary<long, VmValue> _store.
 
-            range = new VmRange((int)_heapPointer, length);
+            range = new VmRange(_heapPointer, length);
             _heapPointer += length;
 
             return true;
